Use long for the bcopy word size instead of int. This is a no-op on the 32-bit systems, but expands well for 64.

+++ sys/powerpc/powerpc/bcopy.c
@@ -52,7 +52,7 @@ __FBSDID("$FreeBSD$");
  * sizeof(word) MUST BE A POWER OF TWO
  * SO THAT wmask BELOW IS ALL ONES
  */
-typedef	int	word;		/* "word" used for optimal copy speed */
+typedef	long	word;		/* "word" used for optimal copy speed */
 
 #define	wsize	sizeof(word)
 #define wmask	(wsize - 1)
